Move tests to grid_test.py
